% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazed.R
\name{downsample}
\alias{downsample}
\title{Downsample Time Series}
\usage{
downsample(data, window_size = 2)
}
\arguments{
\item{data}{The input time series.}

\item{window_size}{The size of the rolling mean window used.}
}
\value{
The downsampled time series.
}
\description{
\code{downsample} samples down a time series with a rolling mean.
}
