% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazed.R
\name{preprocessTs}
\alias{preprocessTs}
\title{Preprocess Time Series for SAZED ensemble}
\usage{
preprocessTs(y)
}
\arguments{
\item{y}{The input time series.}
}
\value{
The detrended and z-normalized time series.
#@examples
season_length <- 26
y <- sin(1:400*2*pi/season_length)
preprocessTs(y)
#@export
}
\description{
\code{preprocessTs} detrends and z-normalizes its argument.
}
